refactor(test): give the chDB SQL shim one owner and unlink test/spec from the heads
The test-support tree had no owner for two shared facts, so each fact was
either copied or reached across a boundary.

The chDB Map-rewrite shim was implemented twice — internal/chclienttest
(the handler-level fake behind the `probe` job) and test/spec (the TXTAR
round-trip lane) each carried their own rewriteMapProjections,
splitOuterSelect, splitProjections, splitAlias, unquoteBackticks,
isMapColumn and Map-column allow-list, plus their own copies of the
seed-DDL backfill kept in sync by a hand-written "keep in lock-step"
comment. The two allow-lists had already diverged in both directions.

Separately, the head-agnostic test/spec runner linked two of three heads:
eval_instant_sweep.go imported internal/promql and traceql_wrap.go
imported internal/traceql, internal/traceql/ast and internal/api/tempo,
while internal/logql was absent. Every head's spec binary therefore
linked every other head's parser, asymmetrically.

Both shims now live in internal/testsql, a substrate-owned package that
depends on nothing. Its allow-list is the union of the two it replaces:
every entry names a column that is Map-typed in the OTel-CH schema on
every lane, so a lane that never projects one simply never matches it.
The rewrite recognises every statement shape either copy handled.

The head coupling is inverted rather than deleted. test/spec gained a
SQLReconstructor option; the traceql lane passes its own wrap
reconstruction in from test/spec/traceqlwrap, and the promql sweep moved
to test/spec/promqlsweep, which the promql lane imports. The shared
runner no longer names a query language.

Evidence: a differential harness ran both retired implementations and the
consolidated one over all 917 fixtures carrying a `-- sql --` section.
The spec pipeline is byte-identical on 916 and the projection count on
917/917; the single delta wraps ExemplarAttributes in a fixture that
declares no `-- seed --`, so it executes nowhere. The handler rewrite
differs on 6, all of them wrapping ResourceAttrs, which the retired
chclienttest list omitted. Both deltas are the union taking effect.

The dependency closure of test/spec under -tags chdb named
internal/promql, internal/traceql, internal/traceql/ast,
internal/api/tempo, internal/api/admit, internal/api/format and
internal/api/httperr; it now names none of them. The logql spec lane's
count of promql/traceql dependencies goes from 2 to 0.

Closes #1824

Closes #1902
