--- internal/chclienttest/rewrite.go
@@ -1,715 +1,0 @@
-//go:build chdb
-
-package chclienttest
-
-import (
-	"slices"
-	"strings"
-)
-
-// chdbEOFSentinel is the spurious "empty row" error chdb-go's parquet
-// driver returns instead of io.EOF at end-of-iteration (see chdb-go
-// v1.11.0 chdb/driver/parquet.go). tolerantRowsErr swallows it so
-// rows.Err() looks normal to callers. Any other error is real.
-const chdbEOFSentinel = "empty row"
-
-func tolerantRowsErr(err error) error {
-	if err == nil {
-		return nil
-	}
-	if strings.Contains(err.Error(), chdbEOFSentinel) {
-		return nil
-	}
-	return err
-}
-
-// mapColumnNames is the conservative allow-list of OTel Map column
-// names that the rewriter will wrap in toJSONString(...) before issuing
-// the query against chDB. We don't have type information here; the
-// rewrite is a textual transform keyed off this list. Extend the list
-// if a new Map column lands in the schema.
-//
-// ExemplarAttributes is the alias the chsql.EmitQueryExemplars outer
-// SELECT projects for `Exemplars.FilteredAttributes` (a Map(LowCardinality
-// (String),String)). Without the toJSONString wrap chDB's parquet driver
-// panics decoding the column as a Go string in the chclienttest scan
-// path — same Map-panic probe Attributes / ResourceAttributes hit.
-// log_attributes / stream_labels are the aliases
-// loki.buildDetectedFieldsSQL projects for `LogAttributes` /
-// `ResourceAttributes` — distinct from the source column names so the
-// toJSONString wrap can't shadow the raw map the WHERE predicate
-// references (CH resolves WHERE identifiers against SELECT aliases
-// first).
-//
-// labels is the alias loki.buildSeriesSQL / buildDetectedLabelsSQL /
-// buildIndexVolumeSQL project for the canonicalised label-set Map
-// (mapSort(`ResourceAttributes`)). Those three are the only `labels`
-// aliases the engine emits and all three are Map-typed, so the textual
-// rewrite is unambiguous. Deliberately distinct from the source column
-// name for the same reason stream_labels is: these queries GROUP BY the
-// alias, and CH resolves GROUP BY identifiers against SELECT aliases
-// before FROM columns.
-var mapColumnNames = []string{
-	"Attributes",
-	"ExemplarAttributes",
-	"LogAttributes",
-	"ResourceAttributes",
-	"ScopeAttributes",
-	"SpanAttributes",
-	"labels",
-	"log_attributes",
-	"stream_labels",
-}
-
-func isMapColumn(name string) bool {
-	for _, c := range mapColumnNames {
-		if name == c {
-			return true
-		}
-	}
-	return false
-}
-
-// rewriteMapProjections wraps any top-level SELECT projection whose
-// alias is a known Map column in toJSONString(...). Only the outermost
-// SELECT is touched — subqueries keep their Map columns raw because
-// CH consumes them server-side.
-//
-// Recognised shapes (mirrors test/spec/runner_chdb.go):
-//
-//	`Attributes`                       → toJSONString(`Attributes`) AS `Attributes`
-//	<expr> AS `Attributes`             → toJSONString(<expr>) AS `Attributes`
-//	`Attributes` AS `Attributes`       → toJSONString(`Attributes`) AS `Attributes`
-//
-// Anything else passes through. If a Map column slips through unwrapped
-// the chdb-go parquet decoder surfaces the cell as NULL against the
-// string scan destination, so the caller fails with "converting NULL to
-// string is unsupported" on the Map column's index.
-func rewriteMapProjections(query string) string {
-	// A leading `WITH …` head — the vector-set-op CSE CTE, the
-	// structural-join `WITH RECURSIVE` closure, or the emitter's
-	// single-evaluation trace-scope binding — sits in front of the outer
-	// SELECT, so the projection split below would not recognise the
-	// statement at all. Peel it, rewrite the SELECT it heads, re-prepend
-	// it verbatim: CTE bodies are subqueries CH consumes server-side, so
-	// their Map columns stay raw like every other subquery's.
-	if head, body := stripWithHead(query); head != "" {
-		return head + rewriteMapProjections(body)
-	}
-	// `SELECT * FROM (<plan>)` — the wrapper the emitter renders around a
-	// plan that needs an outer binding. The star carries the inner
-	// projection's Map column out unwrapped and hides its alias from the
-	// pass below, so rewrite the subquery instead: the star forwards the
-	// inner names and order unchanged, which makes wrapping one level down
-	// produce exactly the same outer wire shape.
-	if inner, ok := starOverSubquery(query); ok {
-		return "SELECT * FROM (" + rewriteMapProjections(inner) + ")"
-	}
-	// Top-level UNION-ALL shape: `(SELECT …) UNION ALL (SELECT …) …`. The
-	// fan-in metadata /series path (internal/api/prom/metadata.go) renders
-	// the combined Sample query as a bare UnionAll of parenthesised SELECT
-	// arms — it does NOT start with `SELECT `, so the single-SELECT rewrite
-	// below would pass it through unwrapped and the Map-typed `Attributes`
-	// column would hit the chdb parquet decoder's NULL path. Rewrite each
-	// arm's outer SELECT independently and re-join.
-	if arms, ok := splitTopLevelUnionAll(query); ok {
-		for i, a := range arms {
-			arms[i] = rewriteMapProjections(a)
-		}
-		return strings.Join(arms, " UNION ALL ")
-	}
-	// A UNION-ALL arm arrives wrapped in its own parens — `(SELECT …)`.
-	// Strip the outer parens, rewrite the inner SELECT, re-wrap.
-	if inner, ok := stripOuterParens(query); ok {
-		return "(" + rewriteMapProjections(inner) + ")"
-	}
-	head, tail := splitOuterSelect(query)
-	if head == "" {
-		return query
-	}
-	projs := splitProjections(head)
-	for i, p := range projs {
-		expr, alias := splitAlias(p)
-		if alias == "" {
-			alias = unquoteBackticks(strings.TrimSpace(expr))
-		}
-		if !isMapColumn(alias) {
-			continue
-		}
-		projs[i] = "toJSONString(" + expr + ") AS `" + alias + "`"
-	}
-	return "SELECT " + strings.Join(projs, ", ") + tail
-}
-
-// stripWithHead peels a leading `WITH <cte-chain> ` off query, returning
-// (head, body) where head is the verbatim prefix up to — and excluding —
-// the outer SELECT, and body is that SELECT. Returns ("", "") when query
-// does not begin with `WITH ` (case-insensitive), so a bare SELECT falls
-// through to the single-SELECT path.
-//
-// The outer SELECT is the first `SELECT ` keyword reached at paren depth
-// 0: every CTE body is parenthesised, whether it is a relational
-// `WITH x AS (SELECT …)`, a `WITH RECURSIVE x AS (…)`, or a scalar
-// `WITH (SELECT …) AS x`, so their own SELECTs sit deeper and are
-// skipped. Mirrors test/spec/runner_chdb.go's peel of the same shapes.
-func stripWithHead(query string) (head, body string) {
-	if !strings.HasPrefix(strings.ToUpper(query), "WITH ") {
-		return "", ""
-	}
-	const sel = "SELECT "
-	depth := 0
-	inStr := byte(0)
-	for i := 0; i < len(query); i++ {
-		c := query[i]
-		switch {
-		case inStr != 0:
-			if c == inStr {
-				inStr = 0
-			}
-			continue
-		case c == '\'' || c == '`':
-			inStr = c
-			continue
-		case c == '(':
-			depth++
-			continue
-		case c == ')':
-			depth--
-			continue
-		}
-		if depth != 0 || i == 0 || i+len(sel) > len(query) {
-			continue
-		}
-		// Standalone keyword only — a `SELECT` that continues an identifier
-		// is not the outer one.
-		if !strings.EqualFold(query[i:i+len(sel)], sel) || !isSQLBreak(query[i-1]) {
-			continue
-		}
-		return query[:i], query[i:]
-	}
-	return "", ""
-}
-
-// isSQLBreak reports whether c can precede a standalone SQL keyword.
-func isSQLBreak(c byte) bool {
-	return c == ' ' || c == '\t' || c == '\n' || c == '\r' || c == ')'
-}
-
-// starOverSubquery matches `SELECT * FROM (<subquery>)` — the wrapper the
-// chsql emitter renders around a plan whose gates reference a binding
-// declared on the outermost statement. Returns the subquery body and
-// ok=true only when the projection is exactly `*` AND the parenthesised
-// FROM operand runs to the end of the statement, so a star over a join, a
-// star followed by ORDER BY / LIMIT, or a star over a bare table all fall
-// through untouched rather than being rewritten on a guess.
-func starOverSubquery(query string) (inner string, ok bool) {
-	head, tail := splitOuterSelect(query)
-	if strings.TrimSpace(head) != "*" {
-		return "", false
-	}
-	return stripOuterParens(tail[len(" FROM "):])
-}
-
-// splitTopLevelUnionAll splits a `<arm> UNION ALL <arm> …` statement on
-// its depth-0 ` UNION ALL ` separators, returning the arms verbatim (each
-// typically a parenthesised `(SELECT …)`). Returns ok=false when no
-// depth-0 ` UNION ALL ` is present, so a plain single SELECT falls through
-// to the single-SELECT rewrite. Single-quoted strings and backtick
-// identifiers shield any ` UNION ALL ` substring inside literals.
-func splitTopLevelUnionAll(query string) (arms []string, ok bool) {
-	const sep = " UNION ALL "
-	var (
-		out   []string
-		start int
-		depth int
-		inStr byte
-	)
-	for i := 0; i < len(query); i++ {
-		c := query[i]
-		switch {
-		case inStr != 0:
-			if c == inStr {
-				inStr = 0
-			}
-		case c == '\'' || c == '`':
-			inStr = c
-		case c == '(':
-			depth++
-		case c == ')':
-			depth--
-		}
-		if depth == 0 && inStr == 0 && i+len(sep) <= len(query) &&
-			strings.EqualFold(query[i:i+len(sep)], sep) {
-			out = append(out, strings.TrimSpace(query[start:i]))
-			i += len(sep) - 1
-			start = i + 1
-		}
-	}
-	if len(out) == 0 {
-		return nil, false
-	}
-	out = append(out, strings.TrimSpace(query[start:]))
-	return out, true
-}
-
-// stripOuterParens returns the contents of a fully-parenthesised
-// expression — `(<inner>)` → `<inner>` — when the leading `(` matches the
-// trailing `)` at depth 0 (i.e. the whole string is one parenthesised
-// group). Returns ok=false otherwise, so a query that merely contains
-// parens (but isn't wholly wrapped) falls through untouched. Quote-aware
-// so a literal `)` inside a string doesn't close the group early.
-func stripOuterParens(s string) (inner string, ok bool) {
-	s = strings.TrimSpace(s)
-	if len(s) < 2 || s[0] != '(' || s[len(s)-1] != ')' {
-		return "", false
-	}
-	depth := 0
-	inStr := byte(0)
-	for i := 0; i < len(s); i++ {
-		c := s[i]
-		switch {
-		case inStr != 0:
-			if c == inStr {
-				inStr = 0
-			}
-		case c == '\'' || c == '`':
-			inStr = c
-		case c == '(':
-			depth++
-		case c == ')':
-			depth--
-			if depth == 0 && i != len(s)-1 {
-				// The opening paren closed before the end — the string is
-				// not a single wrapped group (e.g. `(a) UNION ALL (b)`).
-				return "", false
-			}
-		}
-	}
-	if depth != 0 {
-		return "", false
-	}
-	return strings.TrimSpace(s[1 : len(s)-1]), true
-}
-
-// splitOuterSelect splits `SELECT <projs> FROM ...` at the depth-0
-// FROM. Returns ("", "") if the query isn't a SELECT or the FROM is
-// missing.
-func splitOuterSelect(query string) (head, tail string) {
-	upper := strings.ToUpper(query)
-	if !strings.HasPrefix(upper, "SELECT ") {
-		return "", ""
-	}
-	rest := query[len("SELECT "):]
-	depth := 0
-	for i := 0; i < len(rest); i++ {
-		switch rest[i] {
-		case '(':
-			depth++
-		case ')':
-			depth--
-		}
-		if depth == 0 && i+6 <= len(rest) && strings.EqualFold(rest[i:i+6], " FROM ") {
-			return rest[:i], rest[i:]
-		}
-	}
-	return "", ""
-}
-
-// splitProjections splits a projection list on depth-0 commas.
-// Single-quoted strings and backticks shield commas.
-func splitProjections(s string) []string {
-	var (
-		out   []string
-		buf   strings.Builder
-		depth int
-		inStr byte
-	)
-	for i := 0; i < len(s); i++ {
-		c := s[i]
-		switch {
-		case inStr != 0:
-			if c == inStr {
-				inStr = 0
-			}
-			buf.WriteByte(c)
-		case c == '\'' || c == '`':
-			inStr = c
-			buf.WriteByte(c)
-		case c == '(':
-			depth++
-			buf.WriteByte(c)
-		case c == ')':
-			depth--
-			buf.WriteByte(c)
-		case c == ',' && depth == 0:
-			out = append(out, strings.TrimSpace(buf.String()))
-			buf.Reset()
-		default:
-			buf.WriteByte(c)
-		}
-	}
-	if buf.Len() > 0 {
-		out = append(out, strings.TrimSpace(buf.String()))
-	}
-	return out
-}
-
-// splitAlias splits an `<expr> AS <alias>` projection into (expr,
-// alias). When no AS clause is present returns (s, "").
-func splitAlias(s string) (expr, alias string) {
-	depth := 0
-	inStr := byte(0)
-	lower := strings.ToLower(s)
-	for i := 0; i+4 <= len(s); i++ {
-		c := s[i]
-		switch {
-		case inStr != 0:
-			if c == inStr {
-				inStr = 0
-			}
-		case c == '\'' || c == '`':
-			inStr = c
-		case c == '(':
-			depth++
-		case c == ')':
-			depth--
-		}
-		if depth == 0 && inStr == 0 && lower[i:i+4] == " as " {
-			alias = strings.TrimSpace(s[i+4:])
-			alias = unquoteBackticks(alias)
-			return strings.TrimSpace(s[:i]), alias
-		}
-	}
-	return s, ""
-}
-
-func unquoteBackticks(s string) string {
-	if len(s) >= 2 && s[0] == '`' && s[len(s)-1] == '`' {
-		return s[1 : len(s)-1]
-	}
-	return s
-}
-
-// splitStatements splits a multi-statement SQL script on top-level
-// semicolons. Statements inside single-quoted strings keep their
-// semicolons literal. Mirrors test/spec/runner_chdb.go.
-func splitStatements(s string) []string {
-	var (
-		out   []string
-		buf   strings.Builder
-		inStr bool
-		esc   bool
-	)
-	for i := 0; i < len(s); i++ {
-		c := s[i]
-		switch {
-		case esc:
-			esc = false
-			buf.WriteByte(c)
-		case c == '\\' && inStr:
-			esc = true
-			buf.WriteByte(c)
-		case c == '\'':
-			inStr = !inStr
-			buf.WriteByte(c)
-		case c == ';' && !inStr:
-			out = append(out, buf.String())
-			buf.Reset()
-		default:
-			buf.WriteByte(c)
-		}
-	}
-	if buf.Len() > 0 {
-		out = append(out, buf.String())
-	}
-	return out
-}
-
-func isBlank(s string) bool {
-	return strings.TrimSpace(s) == ""
-}
-
-// metricsTablePrefix scopes the column backfill to OTel-CH metric tables;
-// helper tables a test seeds stay untouched.
-const metricsTablePrefix = "otel_metrics_"
-
-// histogramCountSumTables are the OTel-CH metric tables that store the
-// scalar `Count` / `Sum` companion columns on the same row as the bucket
-// payload — the two tables a `<base>_count` / `<base>_sum` selector
-// lowers against. The gauge and sum tables carry no such columns in
-// production, so the backfill must NOT invent them there: a seed that
-// silently grew a `Count` column would mask a real bug where a query
-// projects a companion column off a table that cannot have one.
-var histogramCountSumTables = []string{
-	"otel_metrics_histogram",
-	"otel_metrics_exponential_histogram",
-}
-
-// backfilledColumn pairs a column name with the definition the backfill
-// injects for it. Every DEFAULT keeps positional INSERTs' value count
-// intact — the backfilled INSERTs carry an explicit column list omitting
-// these columns. A nil `tables` applies the column to every
-// `otel_metrics_*` table; a non-nil one restricts it to those tables, so
-// the injected shape stays faithful to the production DDL per table kind.
-type backfilledColumn struct {
-	name   string
-	ddl    string
-	tables []string
-}
-
-// appliesTo reports whether this column belongs in `table` (already
-// lower-cased by the caller).
-func (c backfilledColumn) appliesTo(table string) bool {
-	if c.tables == nil {
-		return true
-	}
-	return slices.Contains(c.tables, table)
-}
-
-// backfilledColumns is the ordered registry of production OTel-CH metric
-// columns the handler tests' simplified seed DDL may omit. Keep in
-// lock-step with test/spec/roundtrip_prep.go::backfilledColumns.
-//
-// `Count` / `Sum` are here because the classic- and exp-histogram
-// companion lowering projects `toFloat64(Count)` / `toFloat64(Sum)`
-// against those tables whenever a selector can name a `_count` / `_sum`
-// series — including the regex `__name__` fan-out, which reaches BOTH
-// histogram tables for any selector at all. A simplified seed that
-// declares neither column therefore fails at query time with
-// UNKNOWN_IDENTIFIER even though it seeds no histogram rows.
-var backfilledColumns = []backfilledColumn{
-	{name: "ResourceAttributes", ddl: "ResourceAttributes Map(String, String) DEFAULT map()"},
-	{name: "ServiceName", ddl: "ServiceName LowCardinality(String) DEFAULT ''"},
-	{name: "Count", ddl: "Count UInt64 DEFAULT 0", tables: histogramCountSumTables},
-	{name: "Sum", ddl: "Sum Float64 DEFAULT 0", tables: histogramCountSumTables},
-}
-
-// backfillMetricsColumns mirrors the production OTel-CH invariant — every
-// metric table carries the [backfilledColumns] — onto the handler tests'
-// simplified seed DDL, so the read-path projection
-// (`mapConcat(mapUpdate(sanitize(ResourceAttributes), Attributes),
-// map('service_name', ServiceName))`) resolves rather than failing with
-// UNKNOWN_IDENTIFIER. A `CREATE TABLE otel_metrics_*` that declares
-// `Attributes` gets every missing column injected with its DEFAULT, and
-// every subsequent positional `INSERT INTO <that table> VALUES …` is
-// rewritten to an explicit column list (sans the injected columns) so the
-// DEFAULTs fill the gap. Seeds that already declare every column, or
-// already use an explicit INSERT column list, pass through untouched.
-// Mirrors test/spec/roundtrip_prep.go::backfillMetricsColumns.
-func backfillMetricsColumns(stmts []string) []string {
-	cols := map[string][]string{}
-	out := make([]string, 0, len(stmts))
-	for _, stmt := range stmts {
-		if table, names, rewritten, ok := parseMetricsCreate(stmt); ok {
-			cols[table] = names
-			out = append(out, rewritten)
-			continue
-		}
-		if rewritten, ok := rewriteMetricsInsert(stmt, cols); ok {
-			out = append(out, rewritten)
-			continue
-		}
-		out = append(out, stmt)
-	}
-	return out
-}
-
-// createTableHeads are the `CREATE … TABLE ` spellings a seed may open a
-// metric-table definition with. Matching only the bare `CREATE TABLE `
-// form is what let the exotic-PromQL and perf seeds — both written as
-// `CREATE OR REPLACE TABLE` — slip past the backfill and fail at query
-// time with UNKNOWN_IDENTIFIER on a column the read path projects.
-// Longest-first, so `CREATE OR REPLACE TABLE ` is never mis-read as the
-// bare form.
-var createTableHeads = []string{
-	"CREATE OR REPLACE TABLE ",
-	"CREATE TABLE IF NOT EXISTS ",
-	"CREATE TABLE ",
-}
-
-// ddlCreateTableTail returns everything after whichever [createTableHeads]
-// spelling stmt opens with — i.e. the text starting at the table name.
-func ddlCreateTableTail(trimmed string) (string, bool) {
-	upper := strings.ToUpper(trimmed)
-	for _, head := range createTableHeads {
-		if strings.HasPrefix(upper, head) {
-			return trimmed[len(head):], true
-		}
-	}
-	return "", false
-}
-
-func parseMetricsCreate(stmt string) (table string, colNames []string, rewritten string, ok bool) {
-	trimmed := strings.TrimLeft(stmt, " \t\n\r")
-	tail, ok := ddlCreateTableTail(trimmed)
-	if !ok {
-		return "", nil, "", false
-	}
-	name := strings.ToLower(strings.TrimSpace(ddlFirstToken(tail)))
-	if !strings.HasPrefix(name, metricsTablePrefix) {
-		return "", nil, "", false
-	}
-	open := strings.IndexByte(stmt, '(')
-	if open < 0 {
-		return "", nil, "", false
-	}
-	closeParen := ddlMatchParen(stmt, open)
-	if closeParen < 0 {
-		return "", nil, "", false
-	}
-	defs := ddlSplitTopLevelCommas(stmt[open+1 : closeParen])
-	hasAttributes := false
-	declared := make(map[string]bool, len(defs))
-	names := make([]string, 0, len(defs))
-	for _, d := range defs {
-		cn := ddlFirstToken(strings.TrimSpace(d))
-		if cn == "Attributes" {
-			hasAttributes = true
-		}
-		if cn != "" {
-			declared[cn] = true
-			names = append(names, cn)
-		}
-	}
-	missing := make([]string, 0, len(backfilledColumns))
-	for _, c := range backfilledColumns {
-		if !declared[c.name] && c.appliesTo(name) {
-			missing = append(missing, c.ddl)
-		}
-	}
-	if !hasAttributes || len(missing) == 0 {
-		return "", nil, "", false
-	}
-	newDefs := make([]string, 0, len(defs)+len(missing))
-	for _, d := range defs {
-		newDefs = append(newDefs, d)
-		if ddlFirstToken(strings.TrimSpace(d)) != "Attributes" {
-			continue
-		}
-		for _, ddl := range missing {
-			newDefs = append(newDefs, " "+ddl)
-		}
-	}
-	rewritten = stmt[:open+1] + strings.Join(newDefs, ",") + stmt[closeParen:]
-	return name, names, rewritten, true
-}
-
-func rewriteMetricsInsert(stmt string, cols map[string][]string) (string, bool) {
-	trimmed := strings.TrimLeft(stmt, " \t\n\r")
-	prefix := stmt[:len(stmt)-len(trimmed)]
-	upper := strings.ToUpper(trimmed)
-	const needle = "INSERT INTO "
-	if !strings.HasPrefix(upper, needle) {
-		return "", false
-	}
-	rest := trimmed[len(needle):]
-	name := strings.ToLower(strings.TrimSpace(ddlFirstToken(rest)))
-	colNames, tracked := cols[name]
-	if !tracked {
-		return "", false
-	}
-	afterName := strings.TrimLeft(rest[len(ddlFirstToken(rest)):], " \t\n\r")
-	if strings.HasPrefix(afterName, "(") {
-		return "", false // already explicit-column
-	}
-	colList := "(" + strings.Join(colNames, ", ") + ") "
-	return prefix + needle + ddlFirstToken(rest) + " " + colList + afterName, true
-}
-
-func ddlFirstToken(s string) string {
-	s = strings.TrimLeft(s, " \t\n\r")
-	for i := 0; i < len(s); i++ {
-		switch s[i] {
-		case ' ', '\t', '\n', '\r', '(', ',':
-			return s[:i]
-		}
-	}
-	return s
-}
-
-func ddlMatchParen(s string, open int) int {
-	depth := 0
-	inStr := false
-	esc := false
-	for i := open; i < len(s); i++ {
-		c := s[i]
-		switch {
-		case esc:
-			esc = false
-		case c == '\\' && inStr:
-			esc = true
-		case c == '\'':
-			inStr = !inStr
-		case inStr:
-		case c == '(':
-			depth++
-		case c == ')':
-			depth--
-			if depth == 0 {
-				return i
-			}
-		}
-	}
-	return -1
-}
-
-func ddlSplitTopLevelCommas(s string) []string {
-	var (
-		out   []string
-		buf   strings.Builder
-		depth int
-		inStr bool
-		esc   bool
-	)
-	for i := 0; i < len(s); i++ {
-		c := s[i]
-		switch {
-		case esc:
-			esc = false
-			buf.WriteByte(c)
-		case c == '\\' && inStr:
-			esc = true
-			buf.WriteByte(c)
-		case c == '\'':
-			inStr = !inStr
-			buf.WriteByte(c)
-		case inStr:
-			buf.WriteByte(c)
-		case c == '(':
-			depth++
-			buf.WriteByte(c)
-		case c == ')':
-			depth--
-			buf.WriteByte(c)
-		case c == ',' && depth == 0:
-			out = append(out, buf.String())
-			buf.Reset()
-		default:
-			buf.WriteByte(c)
-		}
-	}
-	if strings.TrimSpace(buf.String()) != "" {
-		out = append(out, buf.String())
-	}
-	return out
-}
-
-// promoteCreateTable rewrites a bare `CREATE TABLE …` statement to
-// `CREATE OR REPLACE TABLE …` so re-running a seed against a chDB
-// session that already holds the table is idempotent. Other variants
-// (`CREATE OR REPLACE TABLE`, `CREATE TABLE IF NOT EXISTS`,
-// `CREATE TEMPORARY TABLE`) are left untouched — the rewrite is
-// conservative on purpose. Leading whitespace and comments are
-// preserved verbatim by re-emitting them.
-func promoteCreateTable(stmt string) string {
-	trimmed := strings.TrimLeft(stmt, " \t\n\r")
-	prefix := stmt[:len(stmt)-len(trimmed)]
-	upper := strings.ToUpper(trimmed)
-	// Only touch the bare form. Everything else passes through.
-	const needle = "CREATE TABLE "
-	if !strings.HasPrefix(upper, needle) {
-		return stmt
-	}
-	rest := trimmed[len(needle):]
-	return prefix + "CREATE OR REPLACE TABLE " + rest
-}
